+++ src/main/java/ru/practicum/shareit/booking/BookingStatus.java
@@ -1,0 +1,8 @@
+package ru.practicum.shareit.booking;
+
+public enum BookingStatus {
+    APPROVED,
+    REJECTED,
+    WAITING,
+    CANCELED
+}
