+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -1,0 +1,38 @@
+package ru.practicum.shareit.booking.dto;
+
+import jakarta.validation.constraints.Future;
+import jakarta.validation.constraints.FutureOrPresent;
+import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import ru.practicum.shareit.booking.BookingStatus;
+
+import java.time.LocalDateTime;
+
+@Builder
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode
+public class BookingDto {
+
+    private Integer id;
+
+    @NotNull
+    @FutureOrPresent
+    private LocalDateTime start;
+
+    @NotNull
+    @Future
+    private LocalDateTime end;
+
+    @NotNull
+    private Integer itemId;
+
+    private Integer bookerId;
+
+    private BookingStatus status;
+}
