+++ src/main/java/ru/practicum/shareit/booking/dto/BookingResponseDto.java
@@ -1,0 +1,33 @@
+package ru.practicum.shareit.booking.dto;
+
+
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import ru.practicum.shareit.booking.BookingStatus;
+import ru.practicum.shareit.item.dto.ItemDto;
+import ru.practicum.shareit.user.dto.UserDto;
+
+import java.time.LocalDateTime;
+
+@Builder
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode
+public class BookingResponseDto {
+
+    private Integer id;
+
+    private LocalDateTime start;
+
+    private LocalDateTime end;
+
+    private ItemDto item;
+
+    private UserDto booker;
+
+    private BookingStatus status;
+}
