+++ src/main/java/ru/practicum/shareit/comments/CommentRepository.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.comments;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+
+}
