+++ src/main/java/ru/practicum/shareit/controller/BookingController.java
@@ -17,9 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import ru.practicum.shareit.models.booking.BookingService;
-import ru.practicum.shareit.models.booking.dto.BookingDto;
-import ru.practicum.shareit.models.booking.dto.BookingResponseDto;
+import ru.practicum.shareit.booking.BookingService;
+import ru.practicum.shareit.booking.dto.BookingDto;
+import ru.practicum.shareit.booking.dto.BookingResponseDto;
 
 import java.util.Collection;
 
@@ -43,29 +43,29 @@ public class BookingController {
     }
 
     @GetMapping("/{bookingId}")
-    public BookingResponseDto read(@Positive @RequestHeader(HEADER_USER_ID) Integer userId,
-                                   @Positive @PathVariable Integer bookingId) {
+    public BookingResponseDto getBookingId(@Positive @RequestHeader(HEADER_USER_ID) Integer userId,
+                                           @Positive @PathVariable Integer bookingId) {
         log.trace("Getting booking with id: {} by user with id: {} is started", bookingId, userId);
         return bookingService.findById(userId, bookingId);
     }
 
     @GetMapping
-    public Collection<BookingResponseDto> readUserBookings(@Positive @RequestHeader(HEADER_USER_ID) Integer userId,
+    public Collection<BookingResponseDto> getUserBookings(@Positive @RequestHeader(HEADER_USER_ID) Integer userId,
                                                            @Valid @Pattern(regexp = STATE_REGEX)
                                                            @RequestParam(required = false,
                                                                    defaultValue = "ALL") String state) {
         log.trace("Getting collection of bookings for user-owner with id: {} is started. State is: {}", userId, state);
-        return bookingService.readBookingsForUser(userId, state);
+        return bookingService.getBookingsForUser(userId, state);
     }
 
     @GetMapping("/owner")
-    public Collection<BookingResponseDto> readUserAsOwnerBookings(@Positive @RequestHeader(HEADER_USER_ID) Integer userId,
+    public Collection<BookingResponseDto> getUserAsOwnerBookings(@Positive @RequestHeader(HEADER_USER_ID) Integer userId,
                                                                   @Valid @Pattern(regexp = STATE_REGEX)
                                                                   @RequestParam(required = false,
                                                                           defaultValue = "ALL") String state) {
         log.trace("Getting collection of bookings for user-booker with id: {} is started. State is: {}",
                   userId, state);
-        return bookingService.readBookingsForOwner(userId, state);
+        return bookingService.getBookingsForOwner(userId, state);
     }
 
     @PatchMapping("/{bookingId}")
