+++ src/main/java/ru/practicum/shareit/item/Item.java
@@ -1,0 +1,58 @@
+package ru.practicum.shareit.item;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.BatchSize;
+import ru.practicum.shareit.booking.Booking;
+import ru.practicum.shareit.comments.Comment;
+import ru.practicum.shareit.user.User;
+
+import java.util.Set;
+
+@Builder
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "items", schema = "public")
+public class Item {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column(name = "name")
+    private String name;
+
+    @Column(name = "description")
+    private String description;
+
+    @Column(name = "is_available")
+    private Boolean available;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "owner_id", nullable = false)
+    private User owner;
+
+    @OneToMany(mappedBy = "item", fetch = FetchType.LAZY)
+    @BatchSize(size = 100)
+    private Set<Booking> bookings;
+
+    @OneToMany(mappedBy = "item", fetch = FetchType.LAZY)
+    @BatchSize(size = 100)
+    private Set<Comment> comments;
+}
