--- src/main/java/ru/practicum/shareit/models/booking/BookingStatus.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.models.booking;
-
-public enum BookingStatus {
-    APPROVED,
-    REJECTED,
-    WAITING,
-    CANCELED
-}
