--- src/main/java/ru/practicum/shareit/models/comments/CommentRepository.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.models.comments;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-
-}
