--- src/main/java/ru/practicum/shareit/models/user/UserRepository.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.models.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    Optional<User> findByEmail(String email);
-}
