+++ src/main/java/ru/practicum/shareit/user/UserRepository.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+    Optional<User> findByEmail(String email);
+}
